+++ project/src/components/film-list/film-list.prop.js
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types';
+
+
+export default PropTypes.arrayOf(
+  PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    imgName: PropTypes.string.isRequired,
+    filmTitle: PropTypes.string.isRequired,
+  }));
+
