+++ project/src/mocks/films.js
@@ -1,0 +1,52 @@
+const filmList = [
+  {
+    id: '1',
+    imgName: 'fantastic-beasts-the-crimes-of-grindelwald.jpg',
+    filmTitle: 'Fantastic Beasts: The Crimes of Grindelwald',
+    filmVideo: 'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+  },
+  {
+    id: '2',
+    imgName: 'bohemian-rhapsody.jpg',
+    filmTitle: 'Bohemian Rhapsody',
+    filmVideo: 'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+  },
+  {
+    id: '3',
+    imgName: 'macbeth.jpg',
+    filmTitle: 'Macbeth',
+    filmVideo: 'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+  },
+  {
+    id: '4',
+    imgName: 'aviator.jpg',
+    filmTitle: 'Aviator',
+    filmVideo: 'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+  },
+  {
+    id: '5',
+    imgName: 'we-need-to-talk-about-kevin.jpg',
+    filmTitle: 'We need to talk about Kevin',
+    filmVideo: 'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+  },
+  {
+    id: '6',
+    imgName: 'what-we-do-in-the-shadows.jpg',
+    filmTitle: 'What We Do in the Shadows',
+    filmVideo: 'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+  },
+  {
+    id: '7',
+    imgName: 'revenant.jpg',
+    filmTitle: 'Revenant',
+    filmVideo: 'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+  },
+  {
+    id: '8',
+    imgName: 'johnny-english.jpg',
+    filmTitle: 'Johnny English',
+    filmVideo: 'https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4',
+  },
+];
+
+export default filmList;
